Setup renderHome.js module to create the content of the homepage
The main#content container is now created via js, instead of via
html. Then, the corresponding code in the .html file has been
removed.

The code to do that is in the renderHome.js module. This is
imported and called by an init function in the initWebpage module,
which in turn is called in index.html.