Add main footer + minor code refactoring
The main footer is created, which is different from the credit
footer created before. Indeed, it summarizes the logo, the
contact info, and the opening hours.

This footer is not styled yet.

Some minor code refactoring has been performed, too. Ie, the
resetTab() method has been introduced, to clear the content of a
tab and scrolling back to top, which is used by all the render*()
functions.